chore: move statistical algos into models
